aix: Build SUBSYS.o using $(CC) -r instead of $(LD) -r
This is the only direct use of $(LD), and xlc -r and gcc -r end up with the
same set of symbols and similar performance (noise is high, so hard to say if
equivalent).

Now that $(LD) isn't needed anymore, remove it from src/Makefile.global

While at it, add a comment why -r is used.

Author:
Reviewed-by:
Discussion: https://postgr.es/m/
Backpatch: