Walsender sends unflushed WAL from primary
Walsender is not able to read WAL buffers on primary server and can send
WAL even if it's not written to disk yet. Walsender also sendds
primary's latest flush position so that Walreceiver does not request
write/flush for any WAL further than flush position.

WAL written into buffers on standby server is written to disk when
Walsender asks Walreceiver to do so.